requirements.txt
webapp3
websocket-client
